fixed: warnings

+++ trunk/inetkcp.c
@@ -333,7 +333,7 @@ int ikcp_send(ikcpcb *kcp, const char *buffer, int len)
 	assert(kcp->mss > 0);
 	if (len < 0) return -1;
 
-	if (len <= kcp->mss) count = 1;
+	if (len <= (int)kcp->mss) count = 1;
 	else count = (len + kcp->mss - 1) / kcp->mss;
 
 	if (count > 255) return -2;
